refactor: move bubble into separate file
